ml_project/requirements.txt
hydra-core==1.0.6
pandas~=1.2.4
sklearn~=0.0
scikit-learn~=0.24.1
numpy~=1.20.2
pytest~=6.2.3
Faker~=8.1.1
setuptools~=56.0.0
omegaconf~=2.0.6
typeguard~=2.12.0
marshmallow-dataclass~=8.4.1
flake8~=3.9.1
seaborn~=0.11.1